rent-tracker-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
databases
python-dotenv
passlib[bcrypt]
pydantic
python-jose[cryptography]
alembic